This commits solution to Lab 5: Caching Extents
user@ubuntu:~/git_projects/yfs-git/yfs$ git status
 On branch master
 Changed but not updated:
   (use "git add <file>..." to update what will be committed)
   (use "git checkout -- <file>..." to discard changes in working directory)

   modified:   GNUmakefile
   modified:   extent_client.cc
   modified:   extent_client.h
   modified:   extent_server.cc
   modified:   lock_client_cache.cc
   modified:   lock_client_cache.h
   modified:   yfs_client.cc
   modified:   yfs_client.h

 Untracked files:
   (use "git add <file>..." to include in what will be committed)

   commitmsg
   test.sh

Design
======
Caching at Extent Client
------------------------
In this exercise, we cache extent data/attributes at client end to avoid roundtrips to extent server. We use lock server/client to acquire locks before caching at the extent client side.

get():
If the inode is not present in the hash map, then we invoke load which gets the attributes and data from extent server and populates
struct extent_value {
bool dirty;  // To track puts
std::string data;
extent_protocol::attr ext_attr;
};

put():
If the inode is not present in extent server, we simply cache it in client and finally if the inode is still around, then we dump it back to server on flush() call.

remove():
We remove the inode from hash map

flush():
On release of lock from releaser thread in lock_client_cache, the hook dorelease invokes flush() in extent client. If the data is dirty, we "put" it to server, if not, we "remove" it from the server and delete the entry from client cache if needed.

BUGS FIXED:
1) getattr on extent server was returning OK in all scenarios even if the inode is not in the system
2) Out of Order revoke/retry messages. On lock client, dorelease will be called only "if" the locked has been acquired atleast "once".  doflush flag is set to flag in NONE case when lock state is ACQUIRING. It is then set to true once the state changes to LOCKED state. This ensures that extent_client infact has acquired the lock and so it is valid to call doflush()

test.sh
=======
This script simply does regression test using all test scripts with LOSSY set to 0 and 5.